refactor(cargo-cost-lint): rename module_13 -> budget_config, module_15 -> lint_name_set
Rename the two numbered modules to reflect their contents:
- module_13.rs -> budget_config.rs: holds BudgetConfig, which deserializes
  the [lints] table from budget.toml.
- module_15.rs -> lint_name_set.rs: holds LintNameSet, the HashSet wrapper
  replacing the linear scan over the static lint-name list.

config.rs is kept separate: it holds the legacy  scaffolding
(issue #211/#13, not yet wired into main, no [budget] wrapper), which is a
distinct concern from the  parser in budget_config.rs.

No behaviour changes; the modules remain unwired scaffolding, so the
#[allow(dead_code)] on the mod declarations is retained. The  field
on BudgetConfig is genuinely unused while the module is unwired — left in
place per the refactor guidelines.

Closes #443
